Problems:
1. Download of genbank files was failing.  
2. Fasta file was incomplete as the final sequences were 
   not being added to the outside the while loop.

The corruption of one of the 100 or so ncbi files really forced my hand on 
this script. I have changed this to that we now download the NR fasta file
used by BLAST rather than the genbank translations. This is a single 
file that seems to reliably download and contain roughly the same amount
of sequnece data.

The change in the file format caused me to completely re-write the script 
as the BLAST header lines are a hack on the format, with header lines being 
separated by ctrl-A.  I looked at putting the redundancy back in but this 
lead to a 32 million sequence file, compared to the nonredundant version of 
15 million.  I therefore just take the first GI and move on.

There have been a couple of changes to the schema due to the fact that
secondary accesion and teritary accessions can be 'null' and are much
longer in the NR file. Also, the ncbi_seq table contains primary keys
that are foreign keys in other tables.  These need to be emptied before
deleting the contents of ncbi_seq.