For a while, mongod has been causing issues swapping files in out and from disk. After a little investigation, found that redis will hold the accession map in memory quite happily. This contains the necessary changes to implement Redis as the key/value store.
Ruth also point out that a reference was not being deleted after removal from
the DESC file. This was because it was only going through the database purge
step if that field was populated. In her cases, the removal of the reference
resulted in no references in the DESC file, and so the old incorrect reference
remained in the database. As everything is driven off the database,
the reference made it's way to the release. This issue was also
present for WK and NEST, DBREF lines.