Added code to perform sequence checks of the SEED and the scores.
As part of this, I generated a new Rfam SeqDB object that builds on the Bio::Easel::SqFile.
Like the MySQL database, the object for the sequence database can be obtained from the
config. At the moment, there is only one sequence database, rfamseq. This can be readily
extended. Committing code so that it is safe. 

The SEED sequence QC could be made more efficient, will talk to EPN when he gets back
from lounging around on a beach somewhere on the OBX.