rfsearch.pl: significant change to how search bit score/E-value thresholds are defined.
Previously, DESC BM and SM lines were ignored on purpose because we
were switching from old Infernal 1.0 to Infernal 1.1, which has
different options. Now that all rfsearch's have been rerun with 1.1,
we use the options in BM and SM (which necessarily include an E-value
OR bit score threshold based on how rfsearch defined these options) in
subsequent calls to rfsearch.pl UNLESS the user ignores them
purposefully with the new options -ignorebm and/or -ignoresm. 

If -e or -t are used on the command line, they must be accompanied by
-ignoresm, because as of now all DESC's SM lines have -E or -T
specified. 

The DESC file is now only updated if the search is performed on
RFAMSEQ, so we avoid the problem of overwriting the RFAMSEQ options
with different options used for a one-off search of a genome, for
example. 

rfam.conf: added 'updateDesc' field to seqdb object. If '1' rfsearch
will update DESC file if the corresponding database is used. This will
only be '1' for 'r79rfamseq', the randomized 79-file Rfamseq
database. We don't want SM lines, for example, being updated if we do
a search of a single genome with a search-specific E-value threshold.

rfam_jfrc.conf: Brought up to date (curation, files objects) and added
'updateDesc' field to seqdb, as for rfam.conf.