Added -d option to rfseed.pl to allow deletion of seqs, read either from an outlist file, tblout file, or name list file (one name/start-end per line).
Major reorg of rfseed.pl, most code is now in module functions
(Utils.pm or FamilyIO.pm).

If using an outlist file (or a copy of an outlist file with many lines
removed) for input to rfseed.pl, it must now include the top three
lines of the file (prefixed with "#") because these are used to verify
the format of the file, so that we know which fields are 'start' and
'end'.

rfreplace.pl: some updating to use newly written functions in modules
instead of local functions.