Adding first version of 'create-ss-graphics.pl' script that replaces the view process script RFAMbling.pl from the old codebase. I only reimplemented up to line 133ish of RFAMbling.pl, mostly with Bio-Easel functions (committed separately to that repository) to replace the old Rfam::RfamAlign module functions that were called by RFAMbling.pl.
I put comments in the top of the script explaining how it was changed
from RFAMbling.pl for Rob and/or John to look at as they update the
remainder of the script (SVG, JSON, etc. related code).

I purposefully put this script in top level 'Scripts/' even though I
know it doesn't go there, to encourage whoever knows where it should
go to move it. Will this eventually even be a script? Or a 'plugin'??