Added new configuration, giving a list of files that should be excluded from consideration when committing files to the database. Also added an ordered list of files that need to have their timestamps tweaked after being pulled out of SVN. Added associated methods to the Config.pm module to return the values.
The list of family files was previously stored as a hash in the config, but
then there was a rather convoluted method to convert that hash to an ordered
list of files. I've changed the format of the list in the config so that now
it's stored as an ordered list, and there's no longer any need for the sort
when pulling it out.